网络层,Session层使用了分类切换架构,复杂度分摊到Session切换逻辑上,需要添加一个钩子负责Session类别的切换. 正在开发

+++ src/main/java/org/hashdb/ms/net/exception/UnsupportedProtocolException.java
@@ -11,7 +11,7 @@ import org.hashdb.ms.exception.DBClientException;
  */
 @StandardException
 public class UnsupportedProtocolException extends DBClientException {
-    public static UnsupportedBodyTypeException unsupported(byte unsupportedBodyType) {
+    public static UnsupportedBodyTypeException unsupported(int unsupportedBodyType) {
         return new UnsupportedBodyTypeException("unsupported protocol '" + unsupportedBodyType + "'");
     }
 }
